nifxml.ijs will parse nif.xml and turn that into something representing a file structure
Sadly, xml/sax is in a state of severe neglect, so this uses a very ad-hoc approach to xml parsing. Assumptions include: not much whitespace in certain "legal" places, some assumptions about what elements and attributes are actually present, and so on.

(Using to full generality of xml here would be utterly broken, so don't worry too much about that.)

Still a fair bit of work to do before this is usable because while nif.xml might look like xml it sort of assumes (among other things) we're using python, c++ and a specific qt application to work with it.

So, yeah, not great. But you have to start somewhere...